requirements.txt
django
djangorestframework
django-filter
django-decouple
drf-yasg
psycopg2-binary
Pillow
Celery
